1. Implement validation logic for C++ `Proxy` `validate` method. 2. Hard-code enumeration values for `ValidationMode` to match in both MATLAB and C++. 3. Set default `ValidationMode` to `Minimal` for `arrow.array.ListArrays.fromArrays`. 4. Move `arrow.array.internal.validation.ValidationMode` to `arrow.array.ValidationMode`.
