+++ matlab/src/cpp/arrow/matlab/array/proxy/list_array.cc
@@ -15,10 +15,10 @@
 // specific language governing permissions and limitations
 // under the License.
 
+#include "arrow/matlab/array/validation_mode.h"
 #include "arrow/matlab/array/proxy/list_array.h"
 #include "arrow/matlab/array/proxy/numeric_array.h"
 #include "arrow/matlab/array/proxy/wrap.h"
-#include "arrow/matlab/array/validation/validation_mode.h"
 #include "arrow/matlab/error/error.h"
 #include "arrow/util/utf8.h"
 #include "libmexclass/proxy/ProxyManager.h"
@@ -28,6 +28,7 @@ namespace arrow::matlab::array::proxy {
     ListArray::ListArray(std::shared_ptr<arrow::ListArray> list_array) : proxy::Array{std::move(list_array)} {
         REGISTER_METHOD(ListArray, getValues);
         REGISTER_METHOD(ListArray, getOffsets);
+        REGISTER_METHOD(ListArray, validate);
     }
 
     libmexclass::proxy::MakeResult ListArray::make(const libmexclass::proxy::FunctionArguments& constructor_arguments) {
@@ -105,27 +106,33 @@ namespace arrow::matlab::array::proxy {
 
     void ListArray::validate(libmexclass::proxy::method::Context& context) {
         namespace mda = ::matlab::data;
-        using ValidationMode = array::validation::ValidationMode;
         mda::StructArray args = context.inputs[0];
         const mda::TypedArray<std::uint8_t> validation_mode_mda = args[0]["ValidationMode"];
-        const auto validation_mode_integer = std::uint8_t(validation_mode_mda[0]);
-        const auto validation_mode = static_cast<ValidationMode>(validation_mode_integer); 
+        const auto validation_mode_integer = uint8_t(validation_mode_mda[0]);
         // Convert integer representation to ValidationMode enum.
+        const auto validation_mode = static_cast<ValidationMode>(validation_mode_integer); 
         switch (validation_mode) {
             case ValidationMode::None: {
-                // Do nothing
+                // Do nothing.
                 break;
             }
             case ValidationMode::Minimal: {
-                // Validate
+                MATLAB_ERROR_IF_NOT_OK_WITH_CONTEXT(array->Validate(),
+                        context,
+                        error::ARRAY_VALIDATE_MINIMAL_FAILED);
                 break;
             }
             case ValidationMode::Full: {
-                // ValidateFull
+                MATLAB_ERROR_IF_NOT_OK_WITH_CONTEXT(array->ValidateFull(),
+                        context,
+                        error::ARRAY_VALIDATE_FULL_FAILED);
                 break;
             }
             default: {
-                // Unsupported. Error.
+                // Throw an error if an unsupported enumeration value is provided.
+                const auto msg = "Unsupported ValidationMode enumeration value: " + std::to_string(validation_mode_integer);
+                context.error = libmexclass::error::Error{error::ARRAY_VALIDATE_UNSUPPORTED_ENUM, msg};
+                return;
             } 
         }
     }
