--- matlab/src/matlab/+arrow/+array/+internal/+validation/ValidationMode.m
@@ -1,7 +1,0 @@
-classdef ValidationMode
-    enumeration
-        full,
-        minimal,
-        none
-    end
-end
