+++ matlab/src/matlab/+arrow/+array/ValidationMode.m
@@ -1,0 +1,7 @@
+classdef ValidationMode < uint8
+    enumeration
+        None    (0)
+        Minimal (1)
+        Full    (2)
+    end
+end
